[chop.idx] Make maintenance of the index cross-platform and reliable.

+++ src/Lib/terragear/tg_chopper.hxx
@@ -22,7 +22,7 @@ public:
   }
 
 private:
-    long int GenerateIndex( const std::string& path );
+    uint32_t GenerateIndex( const std::string& path );
     void ClipRow( const tgPolygon& subject, const double& center_lat, const std::string& type );
     tgPolygon Clip( const tgPolygon& subject, const std::string& type, SGBucket& b );
     void Chop( const tgPolygon& subject, const std::string& type );
